refactor: format code with coding standard
