+++ tms-frontend/src/components/DateTimePickerCustom/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 
 import DatePicker from "react-datepicker";
 import { setHours, setMinutes, setSeconds } from "date-fns";
